+++ js/data.js
@@ -1,0 +1,37 @@
+const message = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!',
+];
+const nameAuthor = [
+  'Алексей',
+  'Настя',
+  'Павел',
+  'Денис',
+  'Юля',
+  'Виктория',
+  'Василий',
+  'Денис',
+  'Яна',
+  'Николай',
+  'Евгений',
+  'Дмитрий',
+];
+const description = [
+  'Любовь с первого взгляда',
+  'Осенний лес',
+  'Утром на море',
+  'Разгар рабочего дня у Кекса',
+  'Будущий олимпиец',
+  'Просто обычное фото',
+  'Летом в деревне',
+  'Позитиффф',
+  'Вот оно - счастье!',
+  'На старте',
+  'Первым делом самолёты...',
+  'Бывает жарко)',
+];
+export {message, nameAuthor, description};
